censor emails in usernames
